KSICAPI-268 Removed most of osx warnings. One remainig warning -  ./hash.h:182:2: warning: type specifier missing, defaults to 'int' [-Wimplicit-int] (KSI_FN_DEPRECATED..)

+++ src/ksi/net_file.c
@@ -328,7 +328,7 @@ static int publicationFileReceive(KSI_RequestHandle *handle) {
 	int res;
 	FsClientCtx *fs = NULL;
 	size_t count = 0;
-	size_t size = 0;
+	long int size = 0;
 	unsigned char *buffer = NULL;
 	FILE *f = NULL;
 
@@ -356,7 +356,7 @@ static int publicationFileReceive(KSI_RequestHandle *handle) {
 		goto cleanup;
 	}
 	size = ftell(f);
-	if (size < 0) {
+    if (size < 0) {
 		KSI_pushError(handle->ctx, res = KSI_IO_ERROR, NULL);
 		goto cleanup;
 	}
